Revise id handling.
Track a list of IDs instead of a list of 1s and 0s.  The previous format
could, in a highly unlikely situation, end up with a bunch of unused IDs at
the beginning or in the middle.  For example, if one were to create 2
staircases then delete the 1st one, that ID will never be reused, and
storage would be inefficient (though in a fairly trivial way).

Rewrite how ids are initialized and generated, and how they are handled in
the "Create" function.