BUGFIX: Fix staircase deletion.
The previous version cleared the staircase, but left the old staircase's
spot in place, while removing the ID from the ID list.  The error caused
the code to occasionally try to access a blank staircase.

The fixed version shifts the remaining staircases up to keep the staircases
consistent with the ID list, and leads to cleaner code.